+++ src/services/fetcherHelper/interfaces/ContractHelperInterface.ts
@@ -1,5 +1,0 @@
-import {Contract} from 'ethers';
-
-export interface ContractHelperInterface {
-  getContract(): Contract;
-}