+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Added
 
-## Changed
+## [7.28.0] - 2024-07-02
+### Added
+- Support for `UniswapV3` dex fetcher
+- Migration 7.28.0 adding liquidities and tokens for rootstock
 
 ## [7.27.1] - 2024-06-26
 ### Fixed